Ensure scripts in 'manifest.json' have the correct filepaths
Previously there was some inconsistency between the filepaths in the folder
structure and the `manifest.json` file; the scripts had the `.js` extension in
directory structure example, but `.cocoascript` in `manifest.json`.

This commit ensures that the scripts paths in `manifest.json` match up with the
filepaths according to the folder structure.

Hopefully this will make the guide slightly easier to follow.